Change page.md template to include reminder for changing date
The date should reflect the newest post for the reader.
If the draft date persists, newer (published) posts might appear
below older (published) posts that the reader
has already seen before.

One option is to include a publish script.
Another option is to register all new posts,
and check their dates to ensure an old one is not persisting.
Maybe a simple git hook to save the order of generated posts.