change page header to show metadata succinctly
date will not say 'created'
since the draft version is often created in advance.
it will just show the date.
the last updated date is shown with an arrow, plus the date.
dates are not shown if the post is a draft,
since it's only there because Zola requires it,
and it's changed when published anyway.